requirements.txt
langchain==0.0.111
openai
faiss-cpu
